requirements.txt
flask
pyFirmata
opencv-python
pyserial
ultralytics
torch
trochvision
opencv-python-headless
pyyaml>=5.3
matplotlib

